+++ CHANGES.rst
@@ -19,6 +19,13 @@ To be released.
   as well.
 - Added ``AnnotationError``, ``NoJsonError``, and ``ServiceMethodError``
   exceptions.
+- Since returning a disallowed value which does not match to the return type
+  is the fault the server-side made, the HTTP status code for the case became
+  ``500 Internal Server Error`` instead of ``400 Bad Request``.
+  Also now it writes logs using ``logging`` module.
+- For the mistakes returning ``None`` from a method having non-null return type,
+  now it became to show a more readable and debug-friendly message with a proper
+  response instead of uncaught Python exception.
 - ``WsgiApp.url_map`` attribute was gone.
 - ``/ping/`` resource was gone.